Fix minor issues mentioned in code review
See: https://github.com/mfbutner/CanvasHelpers/pull/10
