+++ SelfAndPeerEvaluations/quiz_results/quiz_reports/json quiz 1.json
@@ -1,1 +1,145 @@
-{"info": {"quiz_id": 181080, "quiz_name": "json quiz 1", "submissions": [296687, 341371, 343238, 343239], "qualitative_subjects": ["Organization", "Communication", "Teamwork", "Attitude", "Contribution of Ideas", "Contribution of Code"]}, "296687": {"name": ["Phan, Christina"], "Organization": [4, 1], "Communication": [1, 1], "Teamwork": [1, 1], "Attitude": [1, 1], "Contribution of Ideas": [1, 1], "Contribution of Code": [1, 1], "valid_solo_submission": [false], "Project Contribution": [0, 0]}, "341371": {"name": ["Bar, Fake"], "Organization": [4], "Communication": [2], "Teamwork": [3], "Attitude": [3], "Contribution of Ideas": [4], "Contribution of Code": [4], "valid_solo_submission": [true], "Project Contribution": [1]}, "343238": {"name": ["SS, AT"], "Organization": [4, 4], "Communication": [4, 4], "Teamwork": [2, 4], "Attitude": [3, 4], "Contribution of Ideas": [1, 3], "Contribution of Code": [3, 2], "valid_solo_submission": [false], "Project Contribution": [0.5, 0.5]}, "343239": {"name": ["AT, SS"], "Organization": [3, 4], "Communication": [3, 3], "Teamwork": [3, 2], "Attitude": [3, 3], "Contribution of Ideas": [2, 2], "Contribution of Code": [4, 4], "valid_solo_submission": [false], "Project Contribution": [0.5, 0.5]}}
+{
+  "info": {
+    "quiz_id": 181080,
+    "quiz_name": "json quiz 1",
+    "submissions": [
+      296687,
+      341371,
+      343238,
+      343239
+    ],
+    "weighting": {
+      "qualitative": 0.4,
+      "project_contribution": 0.6
+    },
+    "qualitative_subjects": [
+      "Organization",
+      "Communication",
+      "Teamwork",
+      "Attitude",
+      "Contribution of Ideas",
+      "Contribution of Code"
+    ]
+  },
+  "296687": {
+    "name": "Phan, Christina",
+    "Organization": [
+      4,
+      1
+    ],
+    "Communication": [
+      1,
+      1
+    ],
+    "Teamwork": [
+      1,
+      1
+    ],
+    "Attitude": [
+      1,
+      1
+    ],
+    "Contribution of Ideas": [
+      1,
+      1
+    ],
+    "Contribution of Code": [
+      1,
+      1
+    ],
+    "valid_solo_submission": false,
+    "Project Contribution": [
+      0,
+      0
+    ]
+  },
+  "341371": {
+    "name": "Bar, Fake",
+    "Organization": [
+      4
+    ],
+    "Communication": [
+      2
+    ],
+    "Teamwork": [
+      3
+    ],
+    "Attitude": [
+      3
+    ],
+    "Contribution of Ideas": [
+      4
+    ],
+    "Contribution of Code": [
+      4
+    ],
+    "valid_solo_submission": true,
+    "Project Contribution": [
+      1
+    ]
+  },
+  "343238": {
+    "name": "SS, AT",
+    "Organization": [
+      4,
+      4
+    ],
+    "Communication": [
+      4,
+      4
+    ],
+    "Teamwork": [
+      2,
+      4
+    ],
+    "Attitude": [
+      3,
+      4
+    ],
+    "Contribution of Ideas": [
+      1,
+      3
+    ],
+    "Contribution of Code": [
+      3,
+      2
+    ],
+    "valid_solo_submission": false,
+    "Project Contribution": [
+      0.5,
+      0.5
+    ]
+  },
+  "343239": {
+    "name": "AT, SS",
+    "Organization": [
+      3,
+      4
+    ],
+    "Communication": [
+      3,
+      3
+    ],
+    "Teamwork": [
+      3,
+      2
+    ],
+    "Attitude": [
+      3,
+      3
+    ],
+    "Contribution of Ideas": [
+      2,
+      2
+    ],
+    "Contribution of Code": [
+      4,
+      4
+    ],
+    "valid_solo_submission": false,
+    "Project Contribution": [
+      0.5,
+      0.5
+    ]
+  }
+}
