--- SelfAndPeerEvaluations/quiz_results/quiz_reports/quiz 1.json
@@ -1,1 +1,0 @@
-{"info": {"quiz_id": 179826, "quiz_name": "quiz 1", "submissions": [343238, 343239, 341371], "qualitative_subjects": ["Organization", "Communication", "Teamwork", "Attitude", "Contribution of Ideas", "Contribution of Code"], "quantitative_overall_subjects": ["Project Contribution"]}, "343238": {"name": ["SS, AT"], "Organization": [4, 4], "Communication": [4, 4], "Teamwork": [4, 4], "Attitude": [4, 4], "Contribution of Ideas": [4, 4], "Contribution of Code": [4, 4], "valid_solo_submission": [false], "Project Contribution": [0.75, 0.5]}, "343239": {"name": ["AT, SS"], "Organization": [3, 3], "Communication": [4, 3], "Teamwork": [4, 3], "Attitude": [4, 3], "Contribution of Ideas": [4, 2], "Contribution of Code": [2, 3], "valid_solo_submission": [false], "Project Contribution": [0.5, 0.25]}, "341371": {"name": ["Bar, Fake"], "Organization": [2], "Communication": [2], "Teamwork": [3], "Attitude": [4], "Contribution of Ideas": [1], "Contribution of Code": [4], "valid_solo_submission": [true], "Project Contribution": [1]}}
