--- SelfAndPeerEvaluations/quiz_results/quiz_reports/quiz 2.json
@@ -1,1 +1,0 @@
-{"info": {"quiz_id": 179827, "quiz_name": "quiz 2", "submissions": [343238], "qualitative_subjects": ["Organization", "Communication", "Teamwork", "Attitude", "Contribution of Ideas", "Contribution of Code"], "quantitative_overall_subjects": ["Project Contribution"]}, "343238": {"name": ["SS, AT"], "Organization": [4, 1], "Communication": [3, 1], "Teamwork": [4, 1], "Attitude": [4, 1], "Contribution of Ideas": [4, 1], "Contribution of Code": [4, 1], "valid_solo_submission": [false], "Project Contribution": [0.75, 0.75]}}
