--- SelfAndPeerEvaluations/quiz_results/quiz_reports/quiz 4.json
@@ -1,1 +1,0 @@
-{"info": {"quiz_id": 179829, "quiz_name": "quiz 4", "submissions": [343238, 343239], "qualitative_subjects": ["Organization", "Communication", "Teamwork", "Attitude", "Contribution of Ideas", "Contribution of Code"], "quantitative_overall_subjects": ["Project Contribution"]}, "343238": {"name": ["SS, AT"], "Organization": [3, 3], "Communication": [3, 3], "Teamwork": [3, 3], "Attitude": [4, 3], "Contribution of Ideas": [2, 3], "Contribution of Code": [2, 3], "valid_solo_submission": [false], "Project Contribution": [0.25, 0.5]}, "343239": {"name": ["AT, SS"], "Organization": [3, 4], "Communication": [3, 4], "Teamwork": [3, 4], "Attitude": [3, 4], "Contribution of Ideas": [2, 4], "Contribution of Code": [3, 4], "valid_solo_submission": [false], "Project Contribution": [0.5, 0.75]}}
